requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohttp==3.8.5
aiomysql==0.2.0
aiosignal==1.3.1
altair==5.3.0
altgraph==0.17.3
annotated-types==0.7.0
appdirs==1.4.4
asgiref==3.8.1
async-timeout==4.0.3
awscli==1.33.40
bcrypt==4.1.2
blinker==1.8.2
boto3==1.34.158
botocore==1.34.158
bs4==0.0.1
CacheControl==0.14.0
cachetools==5.3.3
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==2.2.1
colorama==0.4.6
configparser==5.3.0
contourpy==1.0.7
cramjam==2.8.3
cryptography==42.0.5
cycler==0.11.0
dask==2023.7.0
dateparser==1.1.7
db-dtypes==1.2.0
dill==0.3.8
dnspython==2.4.2
docutils==0.16
dominate==2.9.1
elastic-transport==8.4.0
elasticsearch==7.17.3
email_validator==2.1.1
enchant==0.0.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastparquet==2024.5.0
Fiona==1.9.4.post1
firebase-admin==6.5.0
Flask==3.0.3
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.3
flask-marshmallow==1.2.1
Flask-MySQLdb==2.0.0
flatbuffers==24.3.25
fonttools==4.39.2
frozenlist==1.4.0
fsspec==2023.6.0
gcloud==0.18.3
geopandas==0.13.2
gevent==22.10.2
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-api-python-client==2.135.0
google-auth==2.31.0
google-auth-httplib2==0.2.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
greenlet==2.0.2
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httmock==1.4.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
Hypercorn==0.17.3
hyperframe==6.0.1
influxdb==5.3.1
influxdb-client==1.36.1
ipython-genutils==0.2.0
itsdangerous==2.1.2
jax==0.4.30
jaxlib==0.4.30
jmespath==1.0.1
joblib==1.3.2
jwcrypto==1.5.6
kiwisolver==1.4.4
locket==1.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.2
marshmallow==3.21.3
matplotlib==3.7.1
mdurl==0.1.2
mediapipe==0.10.14
ml-dtypes==0.4.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
mysqlclient==2.2.4
nltk==3.8.1
notebook-as-pdf==0.5.0
numpy==1.24.2
oauth2client==4.1.3
openai==0.28.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.82
opt-einsum==3.3.0
orjson==3.10.3
pandas==1.5.3
paramiko==3.4.0
partd==1.4.0
passlib==1.7.4
Paste==3.5.2
pbr==6.0.0
Pillow==9.4.0
praw==7.7.1
prawcore==2.3.0
priority==2.0.0
prometheus-api-client==0.5.3
proto-plus==1.24.0
protobuf==4.25.3
psycopg2-binary==2.9.9
py4j==0.10.9.7
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycryptodome==3.20.0
pydantic==2.7.3
pydantic_core==2.18.4
pydeck==0.9.1
pyee==8.2.2
Pygments==2.18.0
pyinstaller==5.9.0
pyinstaller-hooks-contrib==2023.1
PyJWT==2.8.0
pymongo==4.5.0
PyMySQL==1.1.0
PyNaCl==1.5.0
pyparsing==3.0.9
PyPDF2==3.0.1
pyppeteer==1.0.2
pyproj==3.6.0
Pyrebase4==4.8.0
pyrsistent==0.19.3
pyspark==3.4.1
pyTelegramBotAPI==4.20.0
python-dotenv==1.0.1
python-jwt==4.1.0
python-multipart==0.0.9
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qiskit==1.1.0
Quart==0.19.6
reactivex==4.0.4
regex==2022.10.31
requests==2.32.3
requests-toolbelt==0.10.1
rich==13.7.1
rsa==4.7.2
rustworkx==0.14.2
s3transfer==0.10.2
scipy==1.13.1
seaborn==0.12.2
shapely==2.0.1
shellingham==1.5.4
smmap==5.0.1
sounddevice==0.4.7
soupsieve==2.4
SQLAlchemy==2.0.30
sshtunnel==0.4.0
starlette==0.37.2
stevedore==5.2.0
streamlit==1.36.0
symengine==0.11.0
sympy==1.12.1
taskgroup==0.0.0a4
tenacity==8.5.0
textblob==0.17.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tqdm==4.65.0
typer==0.12.3
typing_extensions==4.12.1
tzdata==2022.7
tzlocal==4.3
ujson==5.10.0
update-checker==0.18.0
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.30.1
uvloop==0.19.0
vaderSentiment==3.3.2
visitor==0.1.3
waitress==2.1.2
watchdog==4.0.1
watchfiles==0.22.0
webencodings==0.5.1
websockets==10.4
Werkzeug==3.0.3
wsproto==1.2.0
WTForms==3.1.2
yarl==1.9.2
zope.event==4.6
zope.interface==6.0

